moved Copyright from article-api to common package
