+++ article-api/src/main/scala/no/ndla/articleapi/model/domain/Content.scala
@@ -11,6 +11,7 @@ package no.ndla.articleapi.model.domain
 import no.ndla.articleapi.Props
 import no.ndla.common.errors.ValidationException
 import no.ndla.common.model.domain.Availability
+import no.ndla.common.model.domain.article.Copyright
 import org.json4s.{DefaultFormats, FieldSerializer, Formats}
 import org.json4s.FieldSerializer._
 import org.json4s.ext.{EnumNameSerializer, JavaTimeSerializers}
