+++ article-api/src/test/scala/no/ndla/articleapi/service/ConverterServiceTest.scala
@@ -13,6 +13,7 @@ import no.ndla.articleapi.model.api.ImportException
 import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
 import no.ndla.common.model.domain.Availability
+import no.ndla.common.model.domain.article.Copyright
 
 import java.time.LocalDateTime
 import scala.util.Success
