+++ article-api/src/test/scala/no/ndla/articleapi/validation/ContentValidatorTest.scala
@@ -11,6 +11,7 @@ package no.ndla.articleapi.validation
 import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
 import no.ndla.common.errors.{ValidationException, ValidationMessage}
+import no.ndla.common.model.domain.article.Copyright
 import no.ndla.mapping.License.{CC_BY_SA, NA}
 
 import scala.util.Failure
